+++ CHANGES.md
@@ -1,5 +1,14 @@
 # CHANGES
 
+## 0.2
+
+### 0.2.0 - 10-15-2015
+ - Fix:
+    - Added missing 'main' property in package.json
+ - Add:
+    - throwErrors: combine errors into one error object and throw it
+    - whitelist: delete all object properties not validated with 'property'
+
 ## 0.1
 
 ### 0.1.1 - 9-20-15